Model-Pipeline/requirements.txt
# Inherited from data pipeline
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
google-cloud-storage>=2.10.0

# Model pipeline
xgboost>=2.0.0
mlflow>=2.10.0
shap>=0.44.0
scikit-learn>=1.3.0

# Feature engineering / calendar
holidays>=0.46
joblib>=1.3.0

# Hyperparameter optimization (Bayesian search)
optuna>=3.5.0

# Visualizations
matplotlib>=3.8.0

# Drift detection
scipy>=1.11.0